ml/requirements-finetune.txt
# Colab 常为 transformers 5.x；本仓库脚本已兼容 4.x/5.x（eval_strategy、Trainer.processing_class 等）。
# 若需可复现环境，可固定小版本：例如 transformers==5.0.0
transformers>=4.46,<6
datasets>=2.16.0
evaluate>=0.4.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
accelerate>=0.26.0
torch>=2.1.0
